+++ src/components/Facebook.jsx
@@ -46,15 +46,23 @@ const Facebook = () => {
   }, []);
 
   const handleDelete = async (key, url) => {
+    const confirmDelete = window.confirm("Are you sure you want to delete this image?");
+  
+    if (!confirmDelete) return; // Exit the function if the user cancels
+  
     try {
       await remove(ref(db, `facebookPages/${key}`));
+  
+      // Delete the associated image
       const imageRef = storageRef(storage, `Facebook/${url.split('/').pop().split('?')[0]}`);
       await deleteObject(imageRef);
+  
       setImages(images.filter(image => image.key !== key));
     } catch (error) {
       console.error("Error deleting image:", error);
     }
   };
+  
 
   const handleView = (urls, index) => {
     setSelectedImage(urls); // Set an array of URLs
